User policy created

+++ app/controllers/users_controller.rb
@@ -2,11 +2,13 @@ class UsersController < ApplicationController
 
   # GET users/unapproved
   def unapproved_list
+    authorize current_user
     @users = User.where(approved: false)
   end
 
   # POST users/confirm
   def confirm
+    authorize current_user
     user = User.find(params[:id])
     user.approved = true
     respond_to do |format|
@@ -20,6 +22,7 @@ class UsersController < ApplicationController
 
   # DELETE users/reject
   def reject
+    authorize current_user
     user = User.find(params[:id])
     user.destroy
     respond_to do |format|
